bin/notification/requirements.txt
amqp==2.3.2
apns2-client==0.5.4
asn1crypto==0.24.0
billiard==3.5.0.4
celery==4.2.1
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
cryptography==2.3
h2==2.6.2
hpack==3.0.0
hyper==0.7.0
hyperframe==3.2.0
idna==2.7
kombu==4.2.1
pycparser==2.18
pycryptodome==3.6.4
PyJWT==1.6.4
python-http-client==3.1.0
pytz==2018.5
redis==2.10.6
requests==2.19.1
sendgrid==5.4.1
six==1.11.0
ujson==1.35
urllib3==1.23
vine==1.1.4
